make simulation server reusable
